+++ sklearn/linear_model/tests/test_sgd.py
@@ -54,10 +54,8 @@ class SparseSGDRegressor(SGDRegressor):
         X = sp.csr_matrix(X)
         return SGDRegressor.decision_function(self, X, *args, **kw)
 
-##
-## Test Data
-##
 
+# Test Data
 
 # test sample 1
 X = np.array([[-2, -1], [-1, -1], [-1, -2], [1, 1], [1, 2], [2, 1]])
@@ -94,12 +92,46 @@ X5 = np.array([[-2, -1], [-1, -1], [-1, -2], [1, 1], [1, 2], [2, 1]])
 Y5 = [1, 1, 1, 2, 2, 2]
 true_result5 = [0, 1, 1]
 
-##
-## Classification Test Case
-##
 
+# Classification Test Case
 
 class CommonTest(object):
+
+    # a simple implementation of ASGD to use for testing
+    # uses squared loss to find the gradient
+    def asgd(self, X, y, eta, alpha, weight_init=None, intercept_init=0.0):
+        if weight_init is None:
+            weights = np.zeros(X.shape[1])
+        else:
+            weights = weight_init
+
+        average_weights = np.zeros(X.shape[1])
+        intercept = intercept_init
+        average_intercept = 0.0
+        decay = 1.0
+
+        # sparse data has a fixed decay of .01
+        if (isinstance(self, SparseSGDClassifierTestCase) or
+                isinstance(self, SparseSGDRegressorTestCase)):
+            decay = .01
+
+        for i, entry in enumerate(X):
+            p = np.dot(entry, weights)
+            p += intercept
+            gradient = p - y[i]
+            weights *= 1.0 - (eta * alpha)
+            weights += -(eta * gradient * entry)
+            intercept += -(eta * gradient) * decay
+
+            average_weights *= i
+            average_weights += weights
+            average_weights /= i + 1.0
+
+            average_intercept *= i
+            average_intercept += intercept
+            average_intercept /= i + 1.0
+
+        return average_weights, average_intercept
 
     def _test_warm_start(self, X, Y, lr):
         # Test that explicit warm restart...
@@ -113,7 +145,7 @@ class CommonTest(object):
                  coef_init=clf.coef_.copy(),
                  intercept_init=clf.intercept_.copy())
 
-        #... and implicit warm restart are equivalent.
+        # ... and implicit warm restart are equivalent.
         clf3 = self.factory(alpha=0.01, eta0=0.01, n_iter=5, shuffle=False,
                             warm_start=True, learning_rate=lr)
         clf3.fit(X, Y)
@@ -158,6 +190,71 @@ class CommonTest(object):
 
         assert_array_equal(clf.coef_, clf2.coef_)
 
+    def test_plain_has_no_average_attr(self):
+        clf = self.factory(average=True, eta0=.01)
+        clf.fit(X, Y)
+
+        assert_true(hasattr(clf, 'average_coef_'))
+        assert_true(hasattr(clf, 'average_intercept_'))
+        assert_true(hasattr(clf, 'standard_intercept_'))
+        assert_true(hasattr(clf, 'standard_coef_'))
+
+        clf = self.factory()
+        clf.fit(X, Y)
+
+        assert_false(hasattr(clf, 'average_coef_'))
+        assert_false(hasattr(clf, 'average_intercept_'))
+        assert_false(hasattr(clf, 'standard_intercept_'))
+        assert_false(hasattr(clf, 'standard_coef_'))
+
+    def test_late_onset_averaging_not_reached(self):
+        eta0 = .001
+        clf1 = self.factory(average=12, learning_rate="constant",
+                            eta0=eta0, n_iter=2)
+        clf2 = self.factory(learning_rate="constant", eta0=eta0, n_iter=2)
+
+        clf1.fit(X, Y)
+        clf2.fit(X, Y)
+
+        assert_array_almost_equal(clf1.coef_, clf2.coef_)
+        assert_almost_equal(clf1.intercept_, clf1.intercept_)
+
+        clf1 = self.factory(average=13, learning_rate="constant",
+                            eta0=eta0, n_iter=2)
+        clf1.fit(X, Y)
+
+        assert_array_almost_equal(clf1.coef_, clf2.coef_,
+                                  decimal=16)
+        assert_almost_equal(clf1.intercept_, clf2.intercept_,
+                            decimal=16)
+
+    def test_late_onset_averaging_reached(self):
+        eta0 = .001
+        alpha = .0001
+        Y_encode = np.array(Y)
+        Y_encode[Y_encode == 1] = -1.0
+        Y_encode[Y_encode == 2] = 1.0
+
+        clf1 = self.factory(average=7, learning_rate="constant",
+                            loss='squared_loss', eta0=eta0,
+                            alpha=alpha, n_iter=2)
+        clf2 = self.factory(average=0, learning_rate="constant",
+                            loss='squared_loss', eta0=eta0,
+                            alpha=alpha, n_iter=1)
+
+        clf1.fit(X, Y_encode)
+        clf2.fit(X, Y_encode)
+
+        average_weights, average_intercept = \
+            self.asgd(X, Y_encode, eta0, alpha,
+                      weight_init=clf2.coef_.ravel(),
+                      intercept_init=clf2.intercept_)
+
+        assert_array_almost_equal(clf1.coef_.ravel(),
+                                  average_weights.ravel(),
+                                  decimal=16)
+        assert_almost_equal(clf1.intercept_, average_intercept, decimal=16)
+
 
 class DenseSGDClassifierTestCase(unittest.TestCase, CommonTest):
     """Test suite for the dense representation variant of SGD"""
@@ -171,7 +268,7 @@ class DenseSGDClassifierTestCase(unittest.TestCase, CommonTest):
             clf = self.factory(penalty='l2', alpha=0.01, fit_intercept=True,
                                loss=loss, n_iter=10, shuffle=True)
             clf.fit(X, Y)
-            #assert_almost_equal(clf.coef_[0], clf.coef_[1], decimal=7)
+            # assert_almost_equal(clf.coef_[0], clf.coef_[1], decimal=7)
             assert_array_equal(clf.predict(T), true_result)
 
     @raises(ValueError)
@@ -236,6 +333,35 @@ class DenseSGDClassifierTestCase(unittest.TestCase, CommonTest):
         """Checks intercept_ shape for the warm starts in binary case"""
         self.factory().fit(X5, Y5, intercept_init=0)
 
+    def test_average_binary_computed_correctly(self):
+        """Checks the SGDClassifier correctly computes the average weights"""
+        eta = .1
+        alpha = 2.
+        n_samples = 20
+        n_features = 10
+        rng = np.random.RandomState(0)
+        X = rng.normal(size=(n_samples, n_features))
+        w = rng.normal(size=n_features)
+
+        clf = self.factory(loss='squared_loss',
+                           learning_rate='constant',
+                           eta0=eta, alpha=alpha,
+                           fit_intercept=True,
+                           n_iter=1, average=True)
+
+        # simple linear function without noise
+        y = np.dot(X, w)
+        y = np.sign(y)
+
+        clf.fit(X, y)
+
+        average_weights, average_intercept = self.asgd(X, y, eta, alpha)
+        average_weights = average_weights.reshape(1, -1)
+        assert_array_almost_equal(clf.coef_,
+                                  average_weights,
+                                  decimal=14)
+        assert_almost_equal(clf.intercept_, average_intercept, decimal=14)
+
     def test_set_intercept_to_intercept(self):
         """Checks intercept_ shape consistency for the warm starts"""
         # Inconsistent intercept_ shape.
@@ -271,6 +397,29 @@ class DenseSGDClassifierTestCase(unittest.TestCase, CommonTest):
         assert_equal(clf.decision_function([0, 0]).shape, (1, 3))
         pred = clf.predict(T2)
         assert_array_equal(pred, true_result2)
+
+    def test_sgd_multiclass_average(self):
+        eta = .001
+        alpha = .01
+        """Multi-class average test case"""
+        clf = self.factory(loss='squared_loss',
+                           learning_rate='constant',
+                           eta0=eta, alpha=alpha,
+                           fit_intercept=True,
+                           n_iter=1, average=True)
+
+        np_Y2 = np.array(Y2)
+        clf.fit(X2, np_Y2)
+        classes = np.unique(np_Y2)
+
+        for i, cl in enumerate(classes):
+            y_i = np.ones(np_Y2.shape[0])
+            y_i[np_Y2 != cl] = -1
+            average_coef, average_intercept = self.asgd(X2, y_i, eta, alpha)
+            assert_array_almost_equal(average_coef, clf.coef_[i], decimal=16)
+            assert_almost_equal(average_intercept,
+                                clf.intercept_[i],
+                                decimal=16)
 
     def test_sgd_multiclass_with_init_coef(self):
         """Multi-class test case"""
@@ -674,6 +823,84 @@ class DenseSGDRegressorTestCase(unittest.TestCase, CommonTest):
         """Check whether expected ValueError on bad loss"""
         self.factory(loss="foobar")
 
+    def test_sgd_averaged_computed_correctly(self):
+        """Tests the average regressor matches the naive implementation"""
+
+        eta = .001
+        alpha = .01
+        n_samples = 20
+        n_features = 10
+        rng = np.random.RandomState(0)
+        X = rng.normal(size=(n_samples, n_features))
+        w = rng.normal(size=n_features)
+
+        # simple linear function without noise
+        y = np.dot(X, w)
+
+        clf = self.factory(loss='squared_loss',
+                           learning_rate='constant',
+                           eta0=eta, alpha=alpha,
+                           fit_intercept=True,
+                           n_iter=1, average=True)
+
+        clf.fit(X, y)
+        average_weights, average_intercept = self.asgd(X, y, eta, alpha)
+
+        assert_array_almost_equal(clf.coef_,
+                                  average_weights,
+                                  decimal=16)
+        assert_almost_equal(clf.intercept_, average_intercept, decimal=16)
+
+    def test_sgd_averaged_partial_fit(self):
+        """Tests whether the partial fit yields the same average as the fit"""
+        eta = .001
+        alpha = .01
+        n_samples = 20
+        n_features = 10
+        rng = np.random.RandomState(0)
+        X = rng.normal(size=(n_samples, n_features))
+        w = rng.normal(size=n_features)
+
+        # simple linear function without noise
+        y = np.dot(X, w)
+
+        clf = self.factory(loss='squared_loss',
+                           learning_rate='constant',
+                           eta0=eta, alpha=alpha,
+                           fit_intercept=True,
+                           n_iter=1, average=True)
+
+        clf.partial_fit(X[:int(n_samples / 2)][:], y[:int(n_samples / 2)])
+        clf.partial_fit(X[int(n_samples / 2):][:], y[int(n_samples / 2):])
+        average_weights, average_intercept = self.asgd(X, y, eta, alpha)
+
+        assert_array_almost_equal(clf.coef_,
+                                  average_weights,
+                                  decimal=16)
+        assert_almost_equal(clf.intercept_[0], average_intercept, decimal=16)
+
+    def test_average_sparse(self):
+        """Checks the average weights on data with 0s"""
+
+        eta = .001
+        alpha = .01
+        clf = self.factory(loss='squared_loss',
+                           learning_rate='constant',
+                           eta0=eta, alpha=alpha,
+                           fit_intercept=True,
+                           n_iter=1, average=True)
+
+        n_samples = Y3.shape[0]
+
+        clf.partial_fit(X3[:int(n_samples / 2)][:], Y3[:int(n_samples / 2)])
+        clf.partial_fit(X3[int(n_samples / 2):][:], Y3[int(n_samples / 2):])
+        average_weights, average_intercept = self.asgd(X3, Y3, eta, alpha)
+
+        assert_array_almost_equal(clf.coef_,
+                                  average_weights,
+                                  decimal=16)
+        assert_almost_equal(clf.intercept_, average_intercept, decimal=16)
+
     def test_sgd_least_squares_fit(self):
         xmin, xmax = -5, 5
         n_samples = 100
@@ -867,6 +1094,6 @@ def test_underflow_or_overlow():
 
     # model is numerically unstable on unscaled data
     msg_regxp = (r"Floating-point under-/overflow occurred at epoch #.*"
-                  " Scaling input data with StandardScaler or MinMaxScaler"
-                  " might help.")
+                 " Scaling input data with StandardScaler or MinMaxScaler"
+                 " might help.")
     assert_raises_regexp(ValueError, msg_regxp, model.fit, X, y)